refactor: 열거형 수정
