+++ src/main/java/racingcar/model/constants/Phrase.java
@@ -4,15 +4,19 @@ public enum Phrase {
     GAME_START("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)"),
     GAME_TRIAL("시도할 횟수는 몇회인가요?"),
     GAME_RESULT("\n실행 결과"),
-    GAME_WINNER("최종 우승자 : ");
+    GAME_WINNER("최종 우승자 : "),
 
-    private final String phrase;
+    DELIMITER(","),
+    SEPARATOR(" : "),
+    MOVEMENT("-");
 
-    Phrase(String phrase) {
-        this.phrase = phrase;
+    private final String message;
+
+    Phrase(String message) {
+        this.message = message;
     }
 
-    public String getPhrase() {
-        return phrase;
+    public String getMessage() {
+        return message;
     }
 }
