+++ src/main/java/racingcar/model/dto/GameResponseDto.java
@@ -1,8 +1,7 @@
 package racingcar.model.dto;
 
-import static racingcar.model.constants.Rule.MOVEMENT;
-import static racingcar.model.constants.Rule.SEPARATOR;
-import static racingcar.model.constants.Rule.STOP;
+import static racingcar.model.constants.Phrase.MOVEMENT;
+import static racingcar.model.constants.Phrase.SEPARATOR;
 
 import racingcar.model.domain.Car;
 
@@ -18,12 +17,12 @@ public class GameResponseDto {
     public String getCarName() {
         return carName;
     }
-    
+
     public String getGameResult() {
         return String.join(SEPARATOR.getMessage(), carName, convertTo(location));
     }
 
     private static String convertTo(int location) {
-        return String.valueOf(MOVEMENT.getMessage()).repeat(Math.max(STOP.getValue(), location));
+        return String.valueOf(MOVEMENT.getMessage()).repeat(location);
     }
 }
