+++ src/main/java/racingcar/utils/Validator.java
@@ -5,9 +5,9 @@ import static racingcar.model.constants.Error.LENGTH_INVALID;
 import static racingcar.model.constants.Error.NON_DELIMITER;
 import static racingcar.model.constants.Error.NUMBER_NOT_INTEGER;
 import static racingcar.model.constants.Error.NUMBER_RANGE_INVALID;
+import static racingcar.model.constants.Phrase.DELIMITER;
 import static racingcar.model.constants.Rule.CAR_MAX_LENGTH;
 import static racingcar.model.constants.Rule.CAR_MIN_LENGTH;
-import static racingcar.model.constants.Rule.DELIMITER;
 import static racingcar.model.constants.Rule.TRIAL_MIN_LENGTH;
 
 import java.util.List;
